requirements.txt
PyYaml==3.12
requests==2.18.1
beautifulsoup4==4.6.0